ADD : project folder, for multiple projects in same solution/directory
