+++ bom-vizinho-api.sln
@@ -10,7 +10,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		global.json = global.json
 	EndProjectSection
 EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "BomVizinhoAPI", "src\BomVizinhoAPI.xproj", "{BD3A0B30-0BA0-43C7-84BC-CE480DBB6A19}"
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "BomVizinhoAPI", "src\BomVizinhoAPI\BomVizinhoAPI.xproj", "{BD3A0B30-0BA0-43C7-84BC-CE480DBB6A19}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
